Completed up to chapter 2.4
