+++ src/model.ts
@@ -4,20 +4,26 @@ let trackingNumber: string = "FD123455";
 let createDate: Date = new Date();
 let originalCost = 425;
 
+enum InventoryItemType {
+    Computer = "computer",
+    Furniture = "furniture"
+}
+
 interface InventoryItem {
     displayName: string;
-    inventoryType: string;
-    trackingNumber: string;
+    inventoryType: "computer" | "furniture";
+    readonly trackingNumber: string;
     createDate: Date;
-    originalCost: number;
-} 
+    originalCost?: number;
+
+    addNote?: (note: string) => string;
+}
 
 function getInventoryItem(trackingNumber: string): InventoryItem {
     return null;
 }
 
 function saveInventoryItem(item: InventoryItem) {
-
 }
 
 let inventoryItem = getInventoryItem(trackingNumber);
@@ -26,4 +32,9 @@ let updatedInventoryItem = inventoryItem;
 
 inventoryItem.createDate = new Date();
 
-saveInventoryItem(inventoryItem);
+saveInventoryItem({
+    displayName: "MacBook Pro 15 Retina",
+    inventoryType: "computer",
+    trackingNumber: "MBP123456",
+    createDate: new Date(),
+});
