+++ src/pages/addItem.js
@@ -1,7 +1,95 @@
-import inventoryStore from "../inventoryStore.js";
-import inventoryEditor from "./inventoryEditor.js";
+const inventoryEditor = Vue.extend({
+  props: ["item", "category"],
+  render(createEditor) {
+    let editor = null;
 
-export default Vue.extend({
+    switch (this.item.inventoryType) {
+      case "computer":
+        editor = inventoryEditor.computer;
+        break;
+
+      case "furniture":
+        editor = inventoryEditor.furniture;
+        break;
+    }
+
+    return createEditor(editor, { props: this.$props });
+  }
+});
+
+inventoryEditor.computer = Vue.extend({
+  props: ["item"],
+  data: () => ({
+    minYear: 2010,
+    maxYear: new Date().getFullYear()
+  }),
+  template: `
+      <div>
+  
+          <div class="col-sm-3 form-group">
+              <label for="item-brand">Brand</label>
+              <input name="item-brand" type="text" class="form-control" v-model="item.brand" />
+          </div>
+  
+          <div class="col-sm-3 form-group">
+              <label for="item-model">Model</label>
+              <input name="item-model" type="text" class="form-control" v-model="item.model" />
+          </div>
+  
+          <div class="col-sm-3 form-group">
+              <label for="item-year">Year</label>
+              <input name="item-year" type="number":min="minYear" :max="maxYear" class="form-control" v-model="item.year" />
+          </div>
+  
+          <div class="col-sm-3 form-group">
+              <label for="item-serial-number">Serial Number</label>
+              <input name="item-serial-number" type="text" class="form-control" v-model="item.serialNumber" />
+          </div>
+  
+      </div>
+      `
+});
+
+inventoryEditor.furniture = Vue.extend({
+  props: ["item", "category"],
+  computed: {
+    colors() {
+      return this.category.colors;
+    }
+  },
+  template: `
+      <div>
+  
+          <div class="col-sm-6 form-group">
+              <label for="item.manufacturer">Manufacturer</label>
+              <input name="item.manufacturer" type="text" class="form-control" v-model="item.manufacturer" />
+          </div>
+  
+          <div class="col-sm-6 form-group">
+              <label for="item.model">Model / Serial Number / Description</label>
+              <input name="item.model" type="text" class="form-control" v-model="item.model" />
+          </div>
+  
+          <div class="col-sm-3 form-group">
+              <label for="item.material">Material</label>
+              <input name="item.material" type="text" class="form-control" v-model="item.material" />
+          </div>
+  
+          <div v-if="colors" class="col-sm-3 form-group">
+            <label for="item.color">Color</label>
+            <select name="item.color"class="form-control" v-model="item.color">
+              <option disabled value="">-- Select --</option>
+              <option v-for="color in colors" :value="color">
+                {{color}}
+              </option>
+            </select>
+        </div>
+
+      </div>
+      `
+});
+
+const addItemPage = Vue.extend({
   components: { inventoryEditor },
   data: () => ({
     categories: inventoryStore.categories,
@@ -19,18 +107,6 @@ export default Vue.extend({
     },
     subCategories() {
       return this.category.subCategories;
-    },
-    editorComponent() {
-      switch (this.item.inventoryType) {
-        case "computer":
-          return editComputer;
-
-        case "furniture":
-          return editFurniture;
-
-        default:
-          return null;
-      }
     }
   },
   methods: {
@@ -95,7 +171,7 @@ export default Vue.extend({
 
             <div class="col-sm-6 form-group">
                 <label for="item-type">Category</label>
-                <select name="item-type"class="form-control" v-model="item.type">
+                <select name="item-type"class="form-control" v-model="item.inventoryType">
                   <option disabled value="">-- Select --</option>
                   <option v-for="category in categories" :value="category.name">
                     {{category.displayName}}
